[cfe] Move ConstructorBodyBuilderContext
This moves the ConstructorBodyBuilder next to the ConstructorDeclaration and redirects some of its implementation to use the ConstructorDeclaration instead of the SourceConstructorBuilderImpl. This is a step towards supporting constructor patches throught fragments, where the ConstructorBodyBuilderContext must be based on the ConstructorDeclaration rather than the builder.

Change-Id: I978014db5863cadf232227d3f83fd37fd045dfb9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/414502
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>