[dart2wasm] Avoid using deep array copy in places where we only want to copy shallow
The `jsArrayFromDartList` method is a deep array copy: It will call out
to JS to allocate an array and initialize it by calling back to Dart for
each dart list element and `jsify()` each element (which may - if the
element is a list - call `jsArrayFromDartList` again).

For the use cases changed in this CL we know we have a `List<JSAny>` so
there's actually no reason to do this recursively, we only need to
allocate a JS array and fill in the elements from the dart list (without
any processing such as `jsify(0)`ing the elements).

We also add some fast cases to `List<JSAny?>.toJS` for
length=0/1/2/3

Change-Id: Idf82e104a743b883c8af0bd1b3b3234dff6e3d5b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/414780
Reviewed-by: Srujan Gaddam <srujzs@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>