[dart2wasm] Implement `jsifyRaw()` of Dart lists in Dart
Currently `jsifyRaw()` on a Dart list calls out to JS which allocates an
array and then calls back into Dart for each array element, which uses
`jsifyRaw()` on the array elements.

The JS runtimes will not inline the call back into Dart. So there's
actually very little reason to do this in JS, so we can implement it in
Dart.

That in return will also make this more ammenable to tree shaking as the
`$listRead` and `$listLenght` aren't force-exported anymore.

Change-Id: I8b3b7f553ce9fa72880f6ca08edb95b900e98570
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/414781
Reviewed-by: Srujan Gaddam <srujzs@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>