Make Dart2JS use the `y` RegExp flag for `matchAsPrefix`.
The implementation for `matchAsPrefix` predates the sticky `y` flag
for JS regexps, so it had a hack that modified the RegExp pattern to ensure that it only matched at a particular position.
This hack is no longer necessary, so the anchored RegExp can use the
same RegExp source as the plain and global RegExp, which may let the JS engine better reuse RegExp compilation.

Change-Id: I332ebb41cd733346a204345869da16637d775245
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/412420
Auto-Submit: Lasse Nielsen <lrn@google.com>
Commit-Queue: Lasse Nielsen <lrn@google.com>
Reviewed-by: Stephen Adams <sra@google.com>