[dart2wasm] Ensure tests/web/wasm/issue_56234_test runs in unoptimized mode
This test is a regression test for a dart2wasm compiler crash. It tests
that the compiler doesn't crash anymore but the error happens at
runtime.

The particular error that will happen at runtime is that `unreachable`
will be hit (in the place where a unboxed double is attempted to be cast
to an unboxed integer).

Since we run binaryen with `--traps-never-happen` it may optimize this
`unreachable` away and make the test accidentally pass.

So we force this to be run in `-O0` mode, which means we don't run
binaryen on it.

Change-Id: I6460982d4b32d773ee763e9687c8fec0bffa00f0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/415881
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>