Add a migration guide for the analyzer package
Unlike most of our documentation, the snippets in this file are not
compilable, so I've special-cased it in `verify_docs_test.dart`.

We could make the snippets compile, but it would require adding some
significant context, and I'm not sure the context would make the code
easier to read. Let me know if you think we should adhere to the test.

Change-Id: I3af96c378d0a061cdbe069242bd5d73ae09c4205
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/415900
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>