[dart2wasm] Utilize true/false singleton property in `identical()` implementation
We maintain two singleton boxed bool instances for representing `true`
and `false`. So the code for `identical()` no longer has to treat them
as value classes where it has to unbox and compare the WasmI32.

Filed [0] for CFE to improve precision of `ConstantExpression.getStaticType()`.

[0] https://github.com/dart-lang/sdk/issues/60368

Change-Id: I8c02191e974aae9e168b4a826db9a87362949516
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/416941
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>