DAS: In the 'add cast' fix, tidy and take better advantage of promotion
The motivation here was to remove the local duplicate variable,
`target_final`. This variable only existed because the variable it
duplicates, `target`, loses its promoted type inside the
`builder.addDartFileEdit` closure. It loses it's promoted type because
it is multiply assigned.

This CL extracts the first half of the `compute()` method, which is
solely concerned with computing the real "target" and "from type" and
"to type" into a separate method. The diff looks large because of how
the first half of `compute()` is moved into a helper method that is
_below_ the second half of `compute()`.

I think each of these changes then improves readability: extracting
part of a 140-line long method into one helper, and avoiding a local
duplicate variable.

Change-Id: I543bf39e17f0e44011da8741b5f12380e1425f3f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/416883
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>