http profiling: Remove unnecessary async from function
This makes the code simpler, and more readable, and is a no-op.

`_createHttpProfileRequestFromProfileMap` was being always awaited, but
the function did not need to return a Future. So we make it
synchronous, remove the awaits. That then allows
`_getHttpProfileRequest` to not be asynchronous as well.

Change-Id: I1c9854e122e9093e2305f71cc311045f17e81ae9
CoreLibraryReviewExempt: no API; just profiling impl code
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/416260
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>