Add isClosed method to DTD
This allows the Property Editor to check if the DTD connection is still
active and reconnect if not. Because Chrome puts inactive tabs to sleep,
listening to the `done` Future is not sufficient, because it might be
resolved while the tab is asleep.

This will require a release of DTD for us to actually begin using it.

Bug: https://github.com/flutter/devtools/issues/9028
Change-Id: I72e55d0d1d54d3a6d07d34e5e3ac5cde9aca161d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/416841
Reviewed-by: Derek Xu <derekx@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Elliott Brooks <elliottbrooks@google.com>