[cq] Clean up one server test class
This change converts all of the tests to use the `TestCode` approach to
marking positions and ranges, which makes them more consistent and more
readable. It also means that offset-based tests can now be run on the
Windows bot without failure (though this CL doesn't change which tests
are run there and which aren't).

In the process I combined `indexTestUnit` into `_createRefactoring`,
removing one line of code from each of the tests, reducing the visual
clutter.

Change-Id: I02f0d0f46f3965b243abea01828f61a19d276aba
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/417300
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>