analyzer: initialize more fields in ResolverVisitor lazily
I saw that we are initializing a lot of the resolvers/verifiers in
the constructor, but no information from the constructor parameters
is used; they can be initialized in their declaration.

This follows the Effective Dart guideline: "DO initialize fields at
their declaration when possible". I agree it makes the code more
readable.

https://dart.dev/effective-dart/usage#do-initialize-fields-at-their-declaration-when-possible

Change-Id: I3e37f15abff6d185916f84075bc7f6e331f6a255
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/401323
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>