[Beta] Cherry-pick DevTools 2.44.1
Issue description: DevTools cherry-pick to fix a couple issues with the Property Editor: 1) When the computer goes to sleep, the Property Editor is disconnected and does not reconnect when the computer wakes up 2) The Property Editor does not debounce calls to the Analysis Server
What is the fix: 1) Detect when DTD has disconnected, and reload the Property Editor IFRAME 2) Property Editor debounces calls to the Analysis Server at the rate of 600ms
Why cherry-pick: Make sure users trying out the Property Editor on beta are able to use the Property Editor when their computer goes to sleep, and guarantee we are not sending to many requests to the Analysis Server.
Risk: Low, this fix has landed on the main channel of DevTools and has been tested.
Issue link(s): https://github.com/flutter/devtools/issues/9067
Cherry-pick: https://github.com/flutter/devtools/tree/2.44.1
Change-Id: Ia597ef854dc51c94e51e7e2a62c061087dd11eba
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/417903
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Alexander Thomas <athom@google.com>
Reviewed-by: Kenzie Davisson <kenzieschmoll@google.com>