[cfe] Add support for 'call()' methods in dot shorthands.
Any dot shorthand code in the form of `.id()` where `id` is a field or getter should resolve to the `call` method in the declaration of the type of `id`.

Normal static invocations already do this, and this CL adds the extra bit of `call()` resolution to the dot shorthands feature.

Bug: https://github.com/dart-lang/sdk/issues/59758
Change-Id: Ie8a22f251c3c80443e27d9fb307e7ef6b495315e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/418141
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>