[flow analysis] Rename SsaNode.getOrCreatePropertyNode.
The old name (`getProperty`) was unclear because (a) the thing being
returned isn't the property itself but the property's SSA node, and
(b) it didn't give any indication that if the property node was not
present, it would be created.

Thanks to Konstantin for pointing this out.

Change-Id: Ib64315b0bb635545c0ee9253c965277c391f36c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/418661
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>