[analyzer] Fix failing ParameterListTest cases.
These test cases were broken by
https://dart-review.googlesource.com/c/sdk/+/252566 and were marked
with `@failingTest`.

The reason for the breakage was that the aforementioned CL changed the
implementation of `AstComparator.visitSimpleFormalParameter` so that
it compared the parameter's names using `isEqualTokens` rather than
`isEqualNodes`. When this logic is used by `ParameterListTest`,
`isEqualTokens` and `isEqualNodes` are being overridden by methods in
the derived class `ResultComparator`, and one of the special behaviors
of `ResultComparator.isEqualNodes` is to allow the simple identifier
`_k_` to be substituted for a keyword token.

To fix the test, this special behavior needs to be moved from
`ResultComparator.isEqualNodes` to `ResultComparator.isEqualTokens`.

This paves the way for a follow-up CL in which I plan to rewrite
`ResultComparator` so that it doesn't depend on `AstComparator`; that
in turn will allow `AstComparator` to be rmoved.

Change-Id: I8434c7c694e1e5c613df62bf34ee69126fb0c2ec
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/418900
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>